+++ src/test/clojure/clojure/tools/test_logging.clj
@@ -3,31 +3,47 @@
         clojure.tools.logging)
   (:require
     [clojure.set :as set]
-    [clojure.tools.logging.impl :as impl]))
+    [clojure.tools.logging.impl :as impl]
+    [clojure.tools.logging.test :as log-test :refer [logged?
+                                                     matches]]))
 
-(defn test-factory [enabled-set entries-atom agent-used-atom]
-  (let [main-thread (Thread/currentThread)]
-    (reify impl/LoggerFactory
-      (name [_] "test factory")
-      (get-logger [_ log-ns]
-        (reify impl/Logger
-          (enabled? [_ level] (contains? enabled-set level))
-          (write! [_ lvl ex msg]
-            (reset! entries-atom [(str log-ns) lvl ex msg])
-            (reset! agent-used-atom (not (identical? main-thread (Thread/currentThread))))))))))
+; TODO: Improve tools.logging.test so that the ability to check the thread is
+;       not so hacky/out-of-band from the normal flow.
 
-(defmacro with-test-logging
-  [[enabled-level-set log-entry-sym agent-used?-sym] & body]
-  (let [enabled-level-set (or enabled-level-set #{:trace :debug :info :warn :error :fatal})
-        log-entry-sym (or log-entry-sym 'log-entry-sym)
-        agent-used?-sym (or agent-used?-sym 'agent-used?-sym)]
-    `(let [~log-entry-sym (atom nil)
-           ~agent-used?-sym (atom nil)]
-       (binding [*logger-factory* (test-factory
-                                    ~enabled-level-set
-                                    ~log-entry-sym
-                                    ~agent-used?-sym)]
-         ~@body))))
+
+(defn log-entry-with-thread [logger-ns level throwable message]
+  (assoc (log-test/->LogEntry logger-ns level throwable message)
+         ::thread (Thread/currentThread)))
+
+(defn same-thread? [log-entry]
+  (identical? (::thread log-entry)
+              (Thread/currentThread)))
+
+(defn direct-logged? [logger-ns level throwable message]
+  (->> (log-test/matches logger-ns level throwable message)
+    (filter same-thread?)
+    seq
+    boolean))
+
+(defn agent-logged? [logger-ns level throwable message]
+  (->> (log-test/matches logger-ns level throwable message)
+    (remove same-thread?)
+    seq
+    boolean))
+
+(defmacro with-log [& body]
+  `(let [stateful-log# (log-test/atomic-log log-entry-with-thread)
+         logger-factory# (log-test/logger-factory stateful-log# (constantly true))]
+     (binding [log-test/*stateful-log*   stateful-log#
+               *logger-factory* logger-factory#]
+       ~@body)))
+
+(defmacro with-log-level [level & body]
+  `(let [stateful-log# (log-test/atomic-log log-entry-with-thread)
+         logger-factory# (log-test/logger-factory stateful-log# #(= ~level %2))]
+     (binding [log-test/*stateful-log*   stateful-log#
+               *logger-factory* logger-factory#]
+       ~@body)))
 
 (deftest log-yields-nil
   (let [enabled-logger  (reify impl/Logger
@@ -44,14 +60,14 @@
       (is (nil? (log* disabled-logger :debug nil :msg))))))
 
 (deftest log-single-eval
-  (with-test-logging []
+  (with-log
     (let [cnt (atom 0)]
       (log :debug (swap! cnt inc))
       (log :debug (Exception. (str (swap! cnt inc))) (swap! cnt inc))
       (is (== 3 @cnt)))))
 
 (deftest logp-single-eval
-  (with-test-logging []
+  (with-log
     (let [cnt (atom 0)]
       (logp :debug (str (swap! cnt inc)))
       (logp :debug (str (swap! cnt inc)) :foo)
@@ -60,7 +76,7 @@
       (is (== 6 @cnt)))))
 
 (deftest logf-single-eval
-  (with-test-logging []
+  (with-log
     (let [cnt (atom 0)]
       (logf :debug (str (swap! cnt inc)))
       (logf :debug (str (swap! cnt inc)) :foo)
@@ -69,33 +85,30 @@
       (is (== 6 @cnt)))))
 
 (deftest log-msg
-  (with-test-logging [#{:debug} log-entry]
+  (with-log
     (log :debug "foo")
-    (is (= ["clojure.tools.test-logging"
-            :debug
-            nil
-            "foo"]
-           @log-entry))))
+    (is (logged? "clojure.tools.test-logging"
+                 :debug
+                 nil
+                 "foo"))))
 
 (deftest log-ex
-  (with-test-logging [#{:debug} log-entry]
+  (with-log
     (let [e (Exception.)]
       (log :debug e "foo")
-      (is (= ["clojure.tools.test-logging"
-              :debug
-              e
-              "foo"]
-             @log-entry)))))
+      (is (logged? "clojure.tools.test-logging"
+                   :debug
+                   e
+                   "foo")))))
 
 (deftest log-custom-ns
-  (with-test-logging [#{:debug} log-entry]
+  (with-log
     (let [e (Exception.)]
       (log "other.ns" :debug e "foo")
-      (is (= ["other.ns"
-              :debug
-              e
-              "foo"]
-             @log-entry)))))
+      (is (logged? "other.ns"
+                   :debug
+                   e
+                   "foo")))))
 
 (def all-levels #{:trace :debug :info :warn :error :fatal})
 (def default-tx-agent-levels #{:info :warn})
@@ -103,226 +116,196 @@
 
 (deftest log-tx-agent-default
   (doseq [level default-tx-agent-levels]
-    (with-test-logging [#{level} log-entry agent-used?]
+    (with-log-level level
       (dosync
         (log level "foo"))
       (await *logging-agent*)
-      (is (= ["clojure.tools.test-logging" level nil "foo"] @log-entry))
-      (is @agent-used?))))
+      (is (agent-logged? "clojure.tools.test-logging" level nil "foo")))))
 
 (deftest log-no-tx-agent-default
   (doseq [level default-no-tx-agent-levels]
-    (with-test-logging [#{level} log-entry agent-used?]
+    (with-log-level level
       (dosync
         (log level "foo"))
       (await *logging-agent*)
-      (is (= ["clojure.tools.test-logging" level nil "foo"] @log-entry))
-      (is (not @agent-used?)))))
+      (is (direct-logged? "clojure.tools.test-logging" level nil "foo")))))
 
 (deftest log-tx-agent-custom
   (doseq [level default-no-tx-agent-levels]
     (binding [*tx-agent-levels* default-no-tx-agent-levels]
-      (with-test-logging [#{level} log-entry agent-used?]
+      (with-log-level level
         (dosync
           (log level "foo"))
         (await *logging-agent*)
-        (is (= ["clojure.tools.test-logging" level nil "foo"] @log-entry))
-        (is @agent-used?)))))
+        (is (agent-logged? "clojure.tools.test-logging" level nil "foo"))))))
 
 (deftest log-no-tx-agent-custom
   (doseq [level default-tx-agent-levels]
     (binding [*tx-agent-levels* default-no-tx-agent-levels]
-      (with-test-logging [#{level} log-entry agent-used?]
+      (with-log-level level
         (dosync
           (log level "foo"))
         (await *logging-agent*)
-        (is (= ["clojure.tools.test-logging" level nil "foo"] @log-entry))
-        (is (not @agent-used?))))))
+        (is (direct-logged? "clojure.tools.test-logging" level nil "foo"))))))
 
 (deftest log-force-agent
   (doseq [level all-levels]
     (binding [*force* :agent]
-      (with-test-logging [#{level} log-entry agent-used?]
+      (with-log-level level
         (dosync
           (log level "foo"))
         (await *logging-agent*)
-        (is (= ["clojure.tools.test-logging" level nil "foo"] @log-entry))
-        (is @agent-used?)))))
+        (is (agent-logged? "clojure.tools.test-logging" level nil "foo"))))))
 
 (deftest log-force-direct
   (doseq [level all-levels]
     (binding [*force* :direct]
-      (with-test-logging [#{level} log-entry agent-used?]
+      (with-log-level level
         (dosync
           (log level "foo"))
         (await *logging-agent*)
-        (is (= ["clojure.tools.test-logging" level nil "foo"] @log-entry))
-        (is (not @agent-used?))))))
-
-;(comment ; Tests failing on build machine for some unknown reason.
-;         ; See http://groups.google.com/group/clojure-dev/browse_thread/thread/fde32342f3c006bd
-;)
+        (is (direct-logged? "clojure.tools.test-logging" level nil "foo"))))))
 
 (deftest logp-msg-no-optimize
-  (with-test-logging [#{:debug} log-entry]
+  (with-log
     (let [a "foo"
           b "bar"]
       (logp :debug a b))
-    (is (= ["clojure.tools.test-logging" :debug nil "foo bar"] @log-entry))))
+    (is (logged? "clojure.tools.test-logging" :debug nil "foo bar"))))
 
 (deftest logp-msg1
-  (with-test-logging [#{:debug} log-entry]
+  (with-log
     (logp :debug "hello")
-    (is (= ["clojure.tools.test-logging" :debug nil "hello"] @log-entry))))
+    (is (logged? "clojure.tools.test-logging" :debug nil "hello"))))
 
 (deftest logp-msg2
-  (with-test-logging [#{:debug} log-entry]
+  (with-log
     (logp :debug "hello" "world")
-    (is (= ["clojure.tools.test-logging" :debug nil "hello world"]
-           @log-entry))))
+    (is (logged? "clojure.tools.test-logging" :debug nil "hello world"))))
 
 (deftest logp-ex0
-  (with-test-logging [#{:debug} log-entry]
+  (with-log
     (let [e (Exception.)]
       (logp :debug e)
-      (is (= ["clojure.tools.test-logging" :debug nil (print-str e)]
-             @log-entry)))))
+      (is (logged? "clojure.tools.test-logging" :debug nil (print-str e))))))
 
 (deftest logp-ex1
-  (with-test-logging [#{:debug} log-entry]
+  (with-log
     (let [e (Exception.)]
       (logp :debug e "hello")
-      (is (= ["clojure.tools.test-logging" :debug e "hello"]
-             @log-entry)))))
+      (is (logged? "clojure.tools.test-logging" :debug e "hello")))))
 
 (deftest logp-ex2
-  (with-test-logging [#{:debug} log-entry]
+  (with-log
     (let [e (Exception.)]
       (logp :debug e "hello" "world")
-      (is (= ["clojure.tools.test-logging" :debug e "hello world"]
-             @log-entry)))))
+      (is (logged? "clojure.tools.test-logging" :debug e "hello world")))))
 
 (deftest logf-msg-no-optimize
-  (with-test-logging [#{:debug} log-entry]
+  (with-log
     (let [a "foo %s"
           b "bar"]
       (logf :debug a b))
-    (is (= ["clojure.tools.test-logging" :debug nil "foo bar"]
-           @log-entry))))
+    (is (logged? "clojure.tools.test-logging" :debug nil "foo bar"))))
 
 (deftest logf-msg1
-  (with-test-logging [#{:debug} log-entry]
+  (with-log
     (logf :debug "hello")
-    (is (= ["clojure.tools.test-logging" :debug nil "hello"]
-           @log-entry))))
+    (is (logged? "clojure.tools.test-logging" :debug nil "hello"))))
 
 (deftest logf-msg3
-  (with-test-logging [#{:debug} log-entry]
+  (with-log
     (logf :debug "%s %s" "hello" "world")
-    (is (= ["clojure.tools.test-logging" :debug nil "hello world"]
-           @log-entry))))
+    (is (logged? "clojure.tools.test-logging" :debug nil "hello world"))))
 
 (deftest logf-ex0
   (is (thrown? ClassCastException (logf :debug (Exception.)))))
 
 (deftest logf-ex3
-  (with-test-logging [#{:debug} log-entry]
+  (with-log
     (let [e (Exception.)]
       (logf :debug e "%s %s" "hello" "world")
-      (is (= ["clojure.tools.test-logging" :debug e "hello world"]
-             @log-entry)))))
+      (is (logged? "clojure.tools.test-logging" :debug e "hello world")))))
 
 (deftest enabled-true
-  (with-test-logging [#{:fatal}]
+  (with-log-level :fatal
     (is (enabled? :fatal))))
 
 (deftest enabled-false
-  (with-test-logging [#{:debug}]
+  (with-log-level :debug
     (is (not (enabled? :fatal)))))
 
 (deftest spy-default
-  (with-test-logging [#{:debug} log-entry]
+  (with-log-level :debug
     (spy (+ 4 5))
-    (is (= ["clojure.tools.test-logging" :debug nil (format "(+ 4 5)%n=> 9")]
-           @log-entry))))
+    (is (logged? "clojure.tools.test-logging" :debug nil (format "(+ 4 5)%n=> 9")))))
 
 (deftest spy-level
   (doseq [level all-levels]
-    (with-test-logging [#{level} log-entry]
+    (with-log-level level
       (spy level (+ 4 5))
-      (is (= ["clojure.tools.test-logging" level nil (format "(+ 4 5)%n=> 9")]
-             @log-entry)))))
+      (is (logged? "clojure.tools.test-logging" level nil (format "(+ 4 5)%n=> 9"))))))
 
 (deftest spyf-default
-  (with-test-logging [#{:debug} log-entry]
+  (with-log-level :debug
     (spyf "result: %s" (+ 4 5))
-    (is (= ["clojure.tools.test-logging" :debug nil (format "result: 9")]
-           @log-entry))))
+    (is (logged? "clojure.tools.test-logging" :debug nil (format "result: 9")))))
 
 (deftest spyf-level
   (doseq [level all-levels]
-    (with-test-logging [#{level} log-entry]
+    (with-log-level level
       (spyf level "result: %s" (+ 4 5))
-      (is (= ["clojure.tools.test-logging" level nil (format "result: 9")]
-             @log-entry)))))
+      (is (logged? "clojure.tools.test-logging" level nil (format "result: 9"))))))
 
 (deftest capturing
-  (with-test-logging [all-levels log-entry]
+  (with-log
     (log-capture! "foobar")
     (.println System/out "hello world")
-    (is (= ["foobar" :info nil "hello world"]
-           @log-entry))
+    (is (logged? "foobar" :info nil "hello world"))
     (log-uncapture!))
   
-  (with-test-logging [all-levels log-entry]
+  (with-log
     (log-capture! "foobar")
     (.println System/err "oh noes")
-    (is (= ["foobar" :error nil "oh noes"]
-           @log-entry))
+    (is (logged? "foobar" :error nil "oh noes"))
     (log-uncapture!)))
 
 (deftest capturing-level
-  (with-test-logging [all-levels log-entry]
+  (with-log
     (log-capture! "foobar" :error :fatal)
     (.println System/out "hello world")
-    (is (= ["foobar" :error nil "hello world"]
-           @log-entry))
+    (is (logged? "foobar" :error nil "hello world"))
     (log-uncapture!))
   
-  (with-test-logging [all-levels log-entry]
+  (with-log
     (log-capture! "foobar" :error :fatal)
     (.println System/err "oh noes")
-    (is (= ["foobar" :fatal nil "oh noes"]
-           @log-entry))
+    (is (logged? "foobar" :fatal nil "oh noes"))
     (log-uncapture!)))
 
 
 (deftest with-logs-default
-  (with-test-logging [all-levels log-entry]
+  (with-log
     (with-logs "foobar"
       (println "hello world"))
-    (is (= ["foobar" :info nil "hello world"]
-           @log-entry))))
+    (is (logged? "foobar" :info nil "hello world"))))
 
 (deftest with-logs-level
-  (with-test-logging [all-levels log-entry]
+  (with-log
     (with-logs ["foobar" :error :fatal]
       (println "hello world"))
-    (is (= ["foobar" :error nil "hello world"]
-           @log-entry)))
+    (is (logged? "foobar" :error nil "hello world")))
 
-  (with-test-logging [all-levels log-entry]
+  (with-log
     (with-logs ["foobar" :error :fatal]
       (binding [*out* *err*]
         (println "hello world")))
-    (is (= ["foobar" :fatal nil "hello world"]
-           @log-entry))))
+    (is (logged? "foobar" :fatal nil "hello world"))))
 
 (deftest println-style
-  (are [f kw] (with-test-logging [all-levels log-entry]
+  (are [f kw] (with-log
                 (f "hello" "world")
-                (= ["clojure.tools.test-logging" kw nil "hello world"]
-                   @log-entry))
+                (logged? "clojure.tools.test-logging" kw nil "hello world"))
     trace :trace
     debug :debug
     info :info
@@ -332,10 +315,9 @@
 
 (deftest println-style-ex
   (let [e (Exception.)]
-    (are [f kw] (with-test-logging [all-levels log-entry]
+    (are [f kw] (with-log
                   (f e "hello" "world")
-                  (= ["clojure.tools.test-logging" kw e "hello world"]
-                     @log-entry))
+                  (logged? "clojure.tools.test-logging" kw e "hello world"))
       trace :trace
       debug :debug
       info :info
@@ -344,10 +326,9 @@
       fatal :fatal)))
 
 (deftest format-style
-  (are [f kw] (with-test-logging [all-levels log-entry]
-                  (f "%s %s" "hello" "world")
-                  (= ["clojure.tools.test-logging" kw nil "hello world"]
-                     @log-entry))
+  (are [f kw] (with-log
+                (f "%s %s" "hello" "world")
+                (logged? "clojure.tools.test-logging" kw nil "hello world"))
     tracef :trace
     debugf :debug
     infof :info
@@ -357,10 +338,9 @@
 
 (deftest format-style-ex
   (let [e (Exception.)]
-    (are [f kw] (with-test-logging [all-levels log-entry]
+    (are [f kw] (with-log
                   (f e "%s %s" "hello" "world")
-                  (= ["clojure.tools.test-logging" kw e "hello world"]
-                     @log-entry))
+                  (logged? "clojure.tools.test-logging" kw e "hello world"))
       tracef :trace
       debugf :debug
       infof :info