+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ([despite its flaws](https://www.youtube.com/watch?v=oyLBGkS5ICk)).
 
 ## [Unreleased]
+### Added
+- Add support for testing logs in `clojure.tools.logging.test`
+
+### Changed
+- Now passes the context classloader of current thread to `Class/forName` when
+  determining whether logging implementation classes are available on the
+  classpath. This was done to allow testing of the various `impl/*-factory`
+  functions, and seems consistent with clojure internals.
 
 ## [0.4.1] - 2018-05-07
 ### Fixed